@@ -16,7 +16,7 @@ func init() {
 
 type vegaPULS61 struct{}
 
-func (vegaPULS61) NewDevice(c interface{}, rawConf json.RawMessage) map[string]map[string]string {
+func (vegaPULS61) NewDevice(c any, rawConf json.RawMessage) map[string]map[string]string {
 	conn := c.(*connWrap.ConnUtil)
 	var conf struct {
 		DeviceName string `json:"device_name"`
@@ -35,7 +35,7 @@ func (vegaPULS61) NewDevice(c interface{}, rawConf json.RawMessage) map[string]m
 	)
 	var job = func() *float64 {
 		conn.Lock()
-		defer conn.Unlock()
+		defer conn.Unlock() // must be locked to prevent simultaneous operations
 		results, err = client.ReadInputRegisters(2000, 4)
 		if err != nil {
 			global.Log.Error(err)
