@@ -4,7 +4,6 @@ import (
 	"encoding/json"
 	"github.com/google/uuid"
 	"go.uber.org/zap"
-	"io"
 	"sync"
 	"tide/common"
 	"tide/pkg/custype"
@@ -38,21 +37,21 @@ type forwardDataStruct struct {
 	common.DataTimeStruct
 }
 
-func Publish(pub *pubsub.PubSub, msg interface{}, key interface{}) {
+func Publish(pub *pubsub.PubSub, msg any, key any) {
 	if err := pub.Publish(msg, key); err != nil {
 		zap.L().WithOptions(zap.AddCallerSkip(1)).DPanic("publish", zap.Error(err))
 	}
 }
 
-func sendToConfigPubSub(typ string, body interface{}) {
+func sendToConfigPubSub(typ string, body any) {
 	if err := configPubSub.Publish(SendMsgStruct{typ, body}, nil); err != nil {
 		zap.L().WithOptions(zap.AddCallerSkip(1)).DPanic("publish", zap.Error(err))
 	}
 }
 
 type SendMsgStruct struct {
-	Type string      `json:"type"`
-	Body interface{} `json:"body"`
+	Type string `json:"type"`
+	Body any    `json:"body"`
 }
 type RcvMsgStruct struct {
 	Type string          `json:"type"`
@@ -85,22 +84,22 @@ const (
 // connections with downstream or websocket. Used to send data
 var usersConns = struct {
 	sync.Mutex
-	m map[string]map[pubsub.PubConn]int
+	m map[string]map[pubsub.Subscriber]int
 }{
-	m: make(map[string]map[pubsub.PubConn]int),
+	m: make(map[string]map[pubsub.Subscriber]int),
 }
 
-func addUserConn(username string, conn pubsub.PubConn, typ int) {
+func addUserConn(username string, conn pubsub.Subscriber, typ int) {
 	usersConns.Lock()
 	defer usersConns.Unlock()
 	if conns, ok := usersConns.m[username]; ok {
 		conns[conn] = typ
 	} else {
-		usersConns.m[username] = map[pubsub.PubConn]int{conn: typ}
+		usersConns.m[username] = map[pubsub.Subscriber]int{conn: typ}
 	}
 }
 
-func delUserConn(username string, conn pubsub.PubConn) {
+func delUserConn(username string, conn pubsub.Subscriber) {
 	usersConns.Lock()
 	defer usersConns.Unlock()
 	delete(usersConns.m[username], conn)
@@ -112,11 +111,11 @@ func delUserConn(username string, conn pubsub.PubConn) {
 func closeConnByUser(username string, typ int) {
 	usersConns.Lock()
 	defer usersConns.Unlock()
-	conns, ok := usersConns.m[username]
+	subscribers, ok := usersConns.m[username]
 	if ok {
-		for conn, t := range conns {
+		for subscriber, t := range subscribers {
 			if typ&t != 0 {
-				_ = conn.Close()
+				subscriber <- nil
 			}
 		}
 	}
@@ -125,44 +124,154 @@ func closeConnByUser(username string, typ int) {
 func closeAllConn(typ int) {
 	usersConns.Lock()
 	defer usersConns.Unlock()
-	for username, conns := range usersConns.m {
+	for username, subscribers := range usersConns.m {
 		delete(usersConns.m, username)
-		for conn, t := range conns {
+		for subscriber, t := range subscribers {
 			if typ&t != 0 {
-				_ = conn.Close()
+				subscriber <- nil
 			}
 		}
 	}
 }
 
-func changeUserPermissionScope(username string, permissions map[uuid.UUID][]string) {
+func handlePermissionChange(username string, permissions map[uuid.UUID][]string) {
 	usersConns.Lock()
 	defer usersConns.Unlock()
-	conns := usersConns.m[username]
-	if len(conns) > 0 {
-		var permTopic pubsub.TopicMap
-		if permissions != nil {
-			permTopic = uuidStringsMapToTopic(permissions)
+
+	var permTopic pubsub.TopicMap
+	if permissions != nil {
+		permTopic = uuidStringsMapToTopic(permissions)
+	}
+	var syncConfigConns []pubsub.Subscriber
+	for subscriber, t := range usersConns.m[username] {
+		if connTypeWebBrowser&t != 0 {
+			dataPubSub.LimitTopicScope(subscriber, permTopic)
 		}
-		var tmpSyncConns []io.WriteCloser
-		for conn, t := range conns {
-			if connTypeWebBrowser&t != 0 {
-				dataPubSub.LimitTopicScope(conn, permTopic)
+		if connTypeSyncData&t != 0 {
+			subscriber <- nil
+		}
+		if connTypeSyncConfig&t != 0 {
+			syncConfigConns = append(syncConfigConns, subscriber)
+		}
+	}
+	if len(syncConfigConns) > 0 {
+		localAvail, err := db.GetAvailableItems()
+		if err != nil {
+			logger.Error(err.Error())
+			for _, subscriber := range syncConfigConns {
+				subscriber <- nil
 			}
-			if connTypeSyncData&t != 0 {
-				_ = conn.Close()
+			return
+		}
+		var downstreamAvail = make(common.UUIDStringsMap)
+		for _, stationItem := range localAvail {
+			if _, ok := permTopic[stationItem]; ok || permTopic == nil {
+				downstreamAvail[stationItem.StationId] = append(downstreamAvail[stationItem.StationId], stationItem.ItemName)
 			}
+		}
+		j, err := json.Marshal(SendMsgStruct{Type: kMsgUpdateAvailable, Body: downstreamAvail})
+		if err != nil {
+			return
+		}
+		for _, subscriber := range syncConfigConns {
+			select {
+			case subscriber <- j:
+			default:
+				subscriber <- nil
+			}
+		}
+	}
+}
+
+func handleAvailableChange(localAvail map[uuid.UUID][]string) {
+	usersConns.Lock()
+	defer usersConns.Unlock()
+	for username, subscribers := range usersConns.m {
+		var syncConfigSubscribers []pubsub.Subscriber
+		for subscriber, t := range subscribers {
 			if connTypeSyncConfig&t != 0 {
-				tmpSyncConns = append(tmpSyncConns, conn)
+				syncConfigSubscribers = append(syncConfigSubscribers, subscriber)
 			}
 		}
-		if len(tmpSyncConns) > 0 {
-			localAvail, err := db.GetAvailableItems()
+		if len(syncConfigSubscribers) > 0 {
+			user, err := userManager.GetUser(username)
 			if err != nil {
 				logger.Error(err.Error())
-				for _, conn := range tmpSyncConns {
-					_ = conn.Close()
+				for _, subscriber := range syncConfigSubscribers {
+					subscriber <- nil
 				}
+				continue
+			}
+			var permTopic pubsub.TopicMap
+			if user.Role == auth.NormalUser {
+				permissions, err := authorization.GetPermissions(username)
+				if err != nil {
+					logger.Error(err.Error())
+					for _, subscriber := range syncConfigSubscribers {
+						subscriber <- nil
+					}
+					continue
+				}
+				permTopic = uuidStringsMapToTopic(permissions)
+			}
+			var downstreamAvail = make(common.UUIDStringsMap)
+			for stationId, items := range localAvail {
+				for _, itemName := range items {
+					if _, ok := permTopic[common.StationItemStruct{StationId: stationId, ItemName: itemName}]; ok || permTopic == nil {
+						downstreamAvail[stationId] = append(downstreamAvail[stationId], itemName)
+					}
+				}
+			}
+			j, err := json.Marshal(SendMsgStruct{Type: kMsgUpdateAvailable, Body: downstreamAvail})
+			if err != nil {
+				return
+			}
+			for _, subscriber := range syncConfigSubscribers {
+				select {
+				case subscriber <- j:
+				default:
+					subscriber <- nil
+				}
+			}
+		}
+	}
+}
+
+func handleAddItems() {
+	localAvail, err := db.GetAvailableItems()
+	if err != nil {
+		logger.Error(err.Error())
+		return
+	}
+	usersConns.Lock()
+	defer usersConns.Unlock()
+	for username, subscribers := range usersConns.m {
+		var syncConfigSubscribers []pubsub.Subscriber
+		for subscriber, t := range subscribers {
+			if connTypeSyncConfig&t != 0 {
+				syncConfigSubscribers = append(syncConfigSubscribers, subscriber)
+			}
+		}
+		if len(syncConfigSubscribers) > 0 {
+			user, err := userManager.GetUser(username)
+			if err != nil {
+				logger.Error(err.Error())
+				for _, subscriber := range syncConfigSubscribers {
+					subscriber <- nil
+				}
+				continue
+			}
+			var permTopic pubsub.TopicMap
+			if user.Role == auth.NormalUser {
+				permissions, err := authorization.GetPermissions(username)
+				if err != nil {
+					logger.Error(err.Error())
+					for _, subscriber := range syncConfigSubscribers {
+						subscriber <- nil
+					}
+					continue
+				}
+				permTopic = uuidStringsMapToTopic(permissions)
 			}
 			var downstreamAvail = make(common.UUIDStringsMap)
 			for _, stationItem := range localAvail {
@@ -170,61 +279,15 @@ func changeUserPermissionScope(username string, permissions map[uuid.UUID][]stri
 					downstreamAvail[stationItem.StationId] = append(downstreamAvail[stationItem.StationId], stationItem.ItemName)
 				}
 			}
-			for _, conn := range tmpSyncConns {
-				err = json.NewEncoder(conn).Encode(SendMsgStruct{Type: kMsgUpdateAvailable, Body: downstreamAvail})
-				if err != nil {
-					_ = conn.Close()
-				}
+			j, err := json.Marshal(SendMsgStruct{Type: kMsgUpdateAvailable, Body: downstreamAvail})
+			if err != nil {
+				return
 			}
-		}
-	}
-}
-
-func changeUserAvailableScope(localAvail map[uuid.UUID][]string) {
-	usersConns.Lock()
-	defer usersConns.Unlock()
-	for username, conns := range usersConns.m {
-		if len(conns) > 0 {
-			var syncConfigConns []io.WriteCloser
-			for conn, t := range conns {
-				if connTypeSyncConfig&t != 0 {
-					syncConfigConns = append(syncConfigConns, conn)
-				}
-			}
-			if len(syncConfigConns) > 0 {
-				user, err := userManager.GetUser(username)
-				if err != nil {
-					logger.Error(err.Error())
-					for _, conn := range syncConfigConns {
-						_ = conn.Close()
-					}
-					continue
-				}
-				var permTopic pubsub.TopicMap
-				if user.Role == auth.NormalUser {
-					permissions, err := authorization.GetPermissions(username)
-					if err != nil {
-						logger.Error(err.Error())
-						for _, conn := range syncConfigConns {
-							_ = conn.Close()
-						}
-						continue
-					}
-					permTopic = uuidStringsMapToTopic(permissions)
-				}
-				var downstreamAvail = make(common.UUIDStringsMap)
-				for stationId, items := range localAvail {
-					for _, itemName := range items {
-						if _, ok := permTopic[common.StationItemStruct{StationId: stationId, ItemName: itemName}]; ok || permTopic == nil {
-							downstreamAvail[stationId] = append(downstreamAvail[stationId], itemName)
-						}
-					}
-				}
-				for _, conn := range syncConfigConns {
-					err = json.NewEncoder(conn).Encode(SendMsgStruct{Type: kMsgUpdateAvailable, Body: downstreamAvail})
-					if err != nil {
-						_ = conn.Close()
-					}
+			for _, subscriber := range syncConfigSubscribers {
+				select {
+				case subscriber <- j:
+				default:
+					subscriber <- nil
 				}
 			}
 		}
