@@ -5,8 +5,8 @@ import (
 	"database/sql"
 	"encoding/json"
 	"errors"
-	"github.com/Nerzal/gocloak/v11"
-	"github.com/jackc/pgconn"
+	"github.com/Nerzal/gocloak/v13"
+	"github.com/jackc/pgx/v5/pgconn"
 	"net/http"
 	"net/url"
 	"strings"
@@ -16,7 +16,7 @@ import (
 type Keycloak struct {
 	token                          *gocloak.JWT
 	db                             *sql.DB
-	client                         gocloak.GoCloak
+	client                         *gocloak.GoCloak
 	basePath                       string
 	masterUsername, masterPassword string
 	realm, clientId, clientSecret  string
@@ -156,7 +156,8 @@ func (k *Keycloak) AddUser(user auth.User) error {
 	//If the Insert fails in the database, it will return directly
 	_, err = tx.Exec("insert into users(username, role, email, live_camera) values ($1,$2,$3,$4)", user.Username, user.Role, user.Email, user.LiveCamera)
 	if err != nil {
-		if pgErr, ok := err.(*pgconn.PgError); ok && pgErr.Code == "23505" { //ERROR: duplicate key value violates unique constraint "users_username_uindex" (SQLSTATE 23505)
+		var pgErr *pgconn.PgError
+		if errors.As(err, &pgErr) && pgErr.Code == "23505" { //ERROR: duplicate key value violates unique constraint "users_username_uindex" (SQLSTATE 23505)
 			return auth.ErrUserDuplicate
 		}
 		return err
@@ -179,7 +180,8 @@ func (k *Keycloak) AddUser(user auth.User) error {
 
 	userUUID, err := k.client.CreateUser(ctx, k.token.AccessToken, k.realm, kcUser)
 	if err != nil {
-		if apiErr, ok := err.(*gocloak.APIError); !ok || apiErr.Code != http.StatusConflict { // User exists with same username
+		var apiErr *gocloak.APIError
+		if !errors.As(err, &apiErr) || apiErr.Code != http.StatusConflict { // User exists with same username
 			return err
 		}
 		if kcUser, err = k.getKcUserByUsername(ctx, user.Username); err != nil {
@@ -311,7 +313,8 @@ func (k *Keycloak) DelUser(username string) (err error) {
 		//Found a user
 		err = k.client.DeleteUser(ctx, k.token.AccessToken, k.realm, *kcUser.ID)
 		if err != nil {
-			if apiErr, ok := err.(*gocloak.APIError); !ok || apiErr.Code != http.StatusNotFound {
+			var apiErr *gocloak.APIError
+			if !errors.As(err, &apiErr) || apiErr.Code != http.StatusNotFound {
 				return err
 			}
 		}
